test(subagents): split the settle-wake restart-persistent outbox suite
The lane-guard regression tests pushed the settle-wake test file past
the 1000-line lint ceiling; move the restart-persistent outbox
describe block into its own file with the shared harness.
